+++ custom_components/intelbras_amt/alarm_control_panel.py
@@ -42,7 +42,7 @@ async def async_setup_entry(
     async_add_entities: AddEntitiesCallback,
 ) -> None:
     """Set up alarm control panel from a config entry."""
-    coordinator: AMTCoordinator = hass.data[DOMAIN][entry.entry_id]
+    coordinator: AMTCoordinator = hass.data[DOMAIN][entry.entry_id]["coordinator"]
 
     entities: list[AlarmControlPanelEntity] = [
         AMTAlarmControlPanel(coordinator, entry),