+++ custom_components/intelbras_amt/switch.py
@@ -33,7 +33,7 @@ async def async_setup_entry(
     async_add_entities: AddEntitiesCallback,
 ) -> None:
     """Set up switches from a config entry."""
-    coordinator: AMTCoordinator = hass.data[DOMAIN][entry.entry_id]
+    coordinator: AMTCoordinator = hass.data[DOMAIN][entry.entry_id]["coordinator"]
 
     entities: list[SwitchEntity] = []